Add barrier_multi_queue.cpp to Inputs for code reuse
